+++ include/cute/arch/copy_xe_U64.hpp
@@ -1,5 +1,5 @@
 /***************************************************************************************************
- * Copyright (c) 2024 - 2024 Codeplay Software Ltd. All rights reserved.
+ * Copyright (c) 2024 - 2025 Codeplay Software Ltd. All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  *
  * Redistribution and use in source and binary forms, with or without
@@ -30,62 +30,9 @@
  **************************************************************************************************/
 #pragma once
 
-#include <cute/arch/copy.hpp>
-#include <cute/config.hpp>
-#include <cute/arch/xe_copy_4B.hpp>
+#include <cute/util/sycl_vec.hpp>
+#include "cute/config.hpp"
 
-// 64bits No transform Transpose
-SYCL_DEVICE_BUILTIN(
-    cute::intel::ulong __builtin_IB_subgroup_block_read_flat_transpose_u64_k1(
-        intptr_t baseoffset, int width_minus_one, int height_minus_one,
-        int pitch_minus_one, cute::intel::coord_t coord));
-SYCL_DEVICE_BUILTIN(
-    cute::intel::ulong2 __builtin_IB_subgroup_block_read_flat_transpose_u64_k2(
-        intptr_t baseoffset, int width_minus_one, int height_minus_one,
-        int pitch_minus_one, cute::intel::coord_t coord));
-SYCL_DEVICE_BUILTIN(
-    cute::intel::ulong4 __builtin_IB_subgroup_block_read_flat_transpose_u64_k4(
-        intptr_t baseoffset, int width_minus_one, int height_minus_one,
-        int pitch_minus_one, cute::intel::coord_t coord));
-
-
-#if defined(CUTE_ARCH_XE_BUILTIN_ENABLED)
-namespace cute::detail
-{
-template<>
-struct XeSubgroup2DBlockTranspose<8, 1, 8, 1> {
-    template<typename T>
-    CUTE_HOST_DEVICE
-    void operator()(const void* srcBasePointer, int memoryWidth, int memoryHeight, int memoryPitch,
-            cute::intel::coord_t coordinate, T* dstPointer) {
-        *reinterpret_cast<intel::ulong *>(dstPointer) =  __builtin_IB_subgroup_block_read_flat_transpose_u64_k1(
-           reinterpret_cast<long>(srcBasePointer), memoryWidth - 1, memoryHeight - 1, memoryPitch - 1, coordinate);
-    }
-};
-
-template<>
-struct XeSubgroup2DBlockTranspose<8, 2, 8, 1> {
-    template<typename T>
-    CUTE_HOST_DEVICE
-    void operator()(const void* srcBasePointer, int memoryWidth, int memoryHeight, int memoryPitch,
-            cute::intel::coord_t coordinate, T* dstPointer) {
-        *reinterpret_cast<intel::ulong2 *>(dstPointer) =  __builtin_IB_subgroup_block_read_flat_transpose_u64_k2(
-           reinterpret_cast<long>(srcBasePointer), memoryWidth - 1, memoryHeight - 1, memoryPitch - 1, coordinate);
-    }
-};
-
-template<>
-struct XeSubgroup2DBlockTranspose<8, 4, 8, 1> {
-    template<typename T>
-    CUTE_HOST_DEVICE
-    void operator()(const void* srcBasePointer, int memoryWidth, int memoryHeight, int memoryPitch,
-            cute::intel::coord_t coordinate, T* dstPointer) {
-        *reinterpret_cast<intel::ulong4 *>(dstPointer) =  __builtin_IB_subgroup_block_read_flat_transpose_u64_k4(
-           reinterpret_cast<long>(srcBasePointer), memoryWidth - 1, memoryHeight - 1, memoryPitch - 1, coordinate);
-    }
-};
-}
-#endif
 namespace cute
 {
 struct XE_2D_U64x8x1_LD_T {
@@ -95,7 +42,7 @@ struct XE_2D_U64x8x1_LD_T {
   CUTE_HOST_DEVICE static void copy(const void *baseoffset, int width,
                                     int height, int pitch, intel::coord_t coord,
                                     T *dst) {
-#if defined(CUTE_ARCH_XE_ENABLED)
+#if defined(CUTE_ARCH_COPY_XE_ENABLED)
     static_assert(sizeof(T) == 8, "Expected T to have size 8");
     detail::XeSubgroup2DBlockTranspose<8, 1, 8, 1>{}(baseoffset, width, height, pitch, coord, dst);
 #else
@@ -111,7 +58,7 @@ struct XE_2D_U64x8x2_LD_T {
   CUTE_HOST_DEVICE static void copy(const void *baseoffset, int width,
                                     int height, int pitch, intel::coord_t coord,
                                     T *dst) {
-#if defined(CUTE_ARCH_XE_ENABLED)
+#if defined(CUTE_ARCH_COPY_XE_ENABLED)
     static_assert(sizeof(T) == 8, "Expected T to have size 8");
     detail::XeSubgroup2DBlockTranspose<8, 2, 8, 1>{}(baseoffset, width, height, pitch, coord, dst);
 #else
@@ -127,7 +74,7 @@ struct XE_2D_U64x8x4_LD_T {
   CUTE_HOST_DEVICE static void copy(const void *baseoffset, int width,
                                     int height, int pitch, intel::coord_t coord,
                                     T *dst) {
-#if defined(CUTE_ARCH_XE_ENABLED)
+#if defined(CUTE_ARCH_COPY_XE_ENABLED)
     static_assert(sizeof(T) == 8, "Expected T to have size 8");
     detail::XeSubgroup2DBlockTranspose<8, 4, 8, 1>{}(baseoffset, width, height, pitch, coord, dst);
 #else
