+++ include/cute/atom/copy_traits_xe.hpp
@@ -408,9 +408,9 @@ CUTE_HOST_DEVICE constexpr auto make_fragment_layout(TiledCopy &tiled_copy,
   auto [mma_atom_shape, total_mma_atom_iters_M, total_mma_atom_iters_N] = fragment_top_level_shape;
   auto mma_atom_shape_2d = prepend<2>(mma_atom_shape, _1{});
 
-  Int mma_atom_size_M =
+  auto mma_atom_size_M =
       Int<!TiledCopy::is_convention_MN ? size<0>(mma_atom_shape_2d) : size<1>(mma_atom_shape_2d)>{};
-  Int mma_atom_size_N =
+  auto mma_atom_size_N =
       Int<!TiledCopy::is_convention_MN ? size<1>(mma_atom_shape_2d) : size<0>(mma_atom_shape_2d)>{};
 
   using ThreadLayout_ = Shape<_1, Int<size(typename TiledCopy::Traits_LD_t::ThrID{})>>;
@@ -418,15 +418,15 @@ CUTE_HOST_DEVICE constexpr auto make_fragment_layout(TiledCopy &tiled_copy,
                                           ThreadLayout_,
                                           decltype(cute::reverse(ThreadLayout_{}))>;
   auto thread_copy_shape = shape_div(typename TiledCopy::BlockShape{}, ThreadLayout{});
-  Int copy_size_M = size<0>(thread_copy_shape);
-  Int copy_size_N = size<1>(thread_copy_shape);
+  auto copy_size_M = size<0>(thread_copy_shape);
+  auto copy_size_N = size<1>(thread_copy_shape);
 
   static_assert(copy_size_M >= mma_atom_size_M, "MMA atom larger than copy atom is not currently supported.");
   static_assert(copy_size_N >= mma_atom_size_N, "MMA atom larger than copy atom is not currently supported.");
-  Int mma_atom_iters_in_copy_M = copy_size_M / mma_atom_size_M;
-  Int mma_atom_iters_in_copy_N = copy_size_N / mma_atom_size_N;
-  Int copy_iters_M = total_mma_atom_iters_M / mma_atom_iters_in_copy_M;
-  Int copy_iters_N = total_mma_atom_iters_N / mma_atom_iters_in_copy_N;
+  auto mma_atom_iters_in_copy_M = copy_size_M / mma_atom_size_M;
+  auto mma_atom_iters_in_copy_N = copy_size_N / mma_atom_size_N;
+  auto copy_iters_M = total_mma_atom_iters_M / mma_atom_iters_in_copy_M;
+  auto copy_iters_N = total_mma_atom_iters_N / mma_atom_iters_in_copy_N;
 
   auto order = std::conditional_t<TiledCopy::is_convention_MN,
                                   Step<Step<_0, _1>, Step<_2, _4>, Step<_3, _5>>,
@@ -2135,23 +2135,6 @@ struct Copy_Traits<XE_1D_LOAD_GLOBAL<S, D>> {
                              Stride<Int<sizeof_bits<S>::value>, Stride<Int<sizeof_bits<S>::value * 16>, _1>>>;
     // Reference map from (thr,val) to bit
     using RefLayout = DstLayout;
-};
-
-template<class S, class D>
-struct Copy_Traits<PREFETCH<S, D>> {
-    // Logical thread id to thread idx
-    using ThrID = Layout<_16>;
-    // Map from (src-thr,src-val) to bit
-    using SrcLayout = Layout<Shape<_16, Int<sizeof_bits<S>::value>>, Stride<_0, _1>>;
-    // Map from (dst-thr,dst-val) to bit
-    using DstLayout = Layout<Shape <                       _16,Int<sizeof_bits<D>::value>>,
-                             Stride<Int<sizeof_bits<D>::value>,                        _1>>;
-    // Reference map from (thr,val) to bit
-    using RefLayout = DstLayout;
-
-    template <class... CopyArgs>
-    CUTE_HOST_DEVICE
-    Copy_Traits(Copy_Traits<CopyArgs...> const& traits) {}
 };
 
 template<class S, class D>
